Update and rename Heart.Break() MainFrame to Else Heart.Break().Sprak
Refactoring. Made a recursive function that'll find a random room with at least, a single item in it that isn't one of my storage rooms to make it possible to teleport anything away. Perfect for automated defenses in case people with tasers is in the same room as the player?